Separate classes in single-class file #14
Time class in shared
CnvElements in single-class file
