+++ app/lib/graphics/index.ts
@@ -1,19 +1,16 @@
-import { AppSettings } from '@gandolphinnn/shared';
-import { Color } from './src/color';
-import { Coord } from './src/coord';
-import { Circle } from './src/elements';
-import { RenderAction } from './src/types';
-
 export * from './src/angle';
 export * from './src/color';
 export * from './src/coord';
-export * from './src/elements';
 export * from './src/style';
-export * from './src/time';
-export * from './src/image';
 export * from './src/canvas';
 export * from './src/types';
-
-/* export const POINT_DEFAULT = new Circle(Coord.origin, AppSettings.POINT_RADIUS)
-								.setAction(RenderAction.Fill)
-								.setFillStyle(Color.byName('Black')); */
+export * from './src/elements/abstract';
+export * from './src/elements/mesh';
+export * from './src/elements/text';
+export * from './src/elements/line';
+export * from './src/elements/rect';
+export * from './src/elements/poly';
+export * from './src/elements/circle';
+export * from './src/elements/sector';
+export * from './src/elements/slice';
+export * from './src/elements/image';
