+++ app/lib/graphics/src/elements/image.ts
@@ -1,4 +1,4 @@
-import { Coord, MainCanvas, Rect, RenderAction, Size } from '..';
+import { Coord, MainCanvas, Rect, RenderAction, Size } from '../..';
 
 export class Img extends Rect {
