+++ app/lib/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './appSettings';
+export * from './time';
 export * from './types';
